แก้ไข PDA Distribute, PDA Base Close
แก้ไข PDA Distribute, PDA Base Close

+++ snaps.wms.pda/PDA/lib/screens/distribution/models/baseclose_model.dart
@@ -165,7 +165,7 @@ class BaseCloseLine {
   int pv;
   int lv;
   int qtysku;
-  int qtypu;
+  double qtypu;
   double qtyweight;
   double qtyvolume;
   String descalt;
